Move 'set_cell_count_for_project' from tenx_genomics_utils.py to qc/utils.py, and reimplement with CellPlex support.
