+++ auto_process_ngs/tenx_genomics_utils.py
@@ -21,7 +21,6 @@ platforms:
 - cellranger_info
 - spaceranger_info
 - make_qc_summary_html
-- set_cell_count_for_project
 - add_cellranger_args
 """
 
@@ -893,145 +892,6 @@ def make_qc_summary_html(json_file,html_file):
         sample_qc.add(tbl)
     # Write the report
     qc_summary.write(html_file)
-
-def set_cell_count_for_project(project_dir,qc_dir=None):
-    """
-    Set the total number of cells for a project
-
-    Sums the number of cells for each sample in a project
-    (as determined from 'cellranger count' and extracted
-    from the 'metrics_summary.csv' file for scRNA-seq, or
-    from 'cellranger-atac count' and extracted from the
-    'summary.csv' file for scATAC-seq).
-
-    The final count is written to the 'number_of_cells'
-    metadata item for the project.
-
-    Arguments:
-      project_dir (str): path to the project directory
-      qc_dir (str): path to QC directory (if not the default
-        QC directory for the project)
-
-    Returns:
-      Integer: exit code, non-zero values indicate problems
-        were encountered.
-    """
-    # Set up basic info
-    project = AnalysisProject(os.path.basename(project_dir),
-                              project_dir)
-    if qc_dir is None:
-        qc_dir = project.qc_dir
-    qc_dir = os.path.abspath(qc_dir)
-    # Determine which 10x pipeline was used
-    pipeline = None
-    single_cell_platform = project.info.single_cell_platform
-    if single_cell_platform:
-        if single_cell_platform.startswith("10xGenomics Chromium 3'"):
-            pipeline = "cellranger"
-        elif single_cell_platform == "10xGenomics Single Cell ATAC":
-            pipeline = "cellranger-atac"
-        elif single_cell_platform == "10xGenomics Single Cell Multiome":
-            pipeline = "cellranger-arc"
-    if not pipeline:
-        raise NotImplementedError("Not implemented for platform '%s'"
-                                  % single_cell_platform)
-    # Determine possible locations for outputs
-    count_dirs = []
-    # New-style with version and reference data levels
-    qc_info_file = os.path.join(qc_dir,"qc.info")
-    if os.path.exists(qc_info_file):
-        qc_info = AnalysisProjectQCDirInfo(filen=qc_info_file)
-        try:
-            cellranger_refdata = qc_info['cellranger_refdata']
-        except KeyError:
-            cellranger_refdata = None
-        try:
-            cellranger_version = qc_info['cellranger_version']
-        except KeyError:
-            cellranger_version = None
-        if cellranger_version and cellranger_refdata:
-            count_dirs.append(os.path.join(qc_dir,
-                                           "cellranger_count",
-                                           cellranger_version,
-                                           os.path.basename(
-                                               cellranger_refdata)))
-    # Old-style without additional subdirectories
-    count_dirs.append(os.path.join(qc_dir,
-                                   "cellranger_count"))
-    # Check each putative output location in turn
-    for count_dir in count_dirs:
-        # Check that the directory exists
-        if not os.path.exists(count_dir):
-            logger.warning("%s: not found" % count_dir)
-            continue
-        print("Looking for outputs under %s" % count_dir)
-        # Loop over samples and collect cell numbers for each
-        number_of_cells = 0
-        try:
-            for sample in project.samples:
-                print("- %s" % sample)
-                outs_dir = os.path.join(count_dir,
-                                        sample.name,
-                                        "outs")
-                if pipeline == "cellranger":
-                    # Single cell/nuclei RNA-seq output
-                    metrics_summary_csv = os.path.join(
-                        outs_dir,
-                        "metrics_summary.csv")
-                    if not os.path.exists(metrics_summary_csv):
-                        raise Exception("Failed to add cell count "
-                                        "for sample '%s': missing "
-                                        "file '%s'" %
-                                        (sample.name,
-                                         metrics_summary_csv))
-                    # Extract cell numbers
-                    metrics = GexSummary(metrics_summary_csv)
-                    number_of_cells += \
-                        metrics.estimated_number_of_cells
-                elif pipeline == "cellranger-atac":
-                    # Single cell ATAC-seq output
-                    summary_csv = os.path.join(
-                        outs_dir,
-                        "summary.csv")
-                    if not os.path.exists(summary_csv):
-                        raise Exception("Failed to add cell count "
-                                        "for sample '%s': missing "
-                                        "file '%s'" %
-                                        (sample.name,
-                                         summary_csv))
-                    # Extract cell numbers
-                    try:
-                        # Cellranger ATAC pre-2.0.0 uses 'annotated_cells'
-                        number_of_cells += AtacSummary(summary_csv).\
-                                           annotated_cells
-                    except AttributeError:
-                        # Cellranger ATAC 2.0.0 uses 'Estimated number of
-                        # cells'
-                        number_of_cells += AtacSummary(summary_csv).\
-                                           estimated_number_of_cells
-                elif pipeline == "cellranger-arc":
-                    # Single cell multiome output
-                    summary_csv = os.path.join(
-                        outs_dir,
-                        "summary.csv")
-                    if not os.path.exists(summary_csv):
-                        raise Exception("Failed to add cell count "
-                                        "for sample '%s': missing "
-                                        "file '%s'" %
-                                        (sample.name,
-                                         summary_csv))
-                    # Extract cell numbers
-                    number_of_cells += MultiomeSummary(summary_csv).\
-                                       estimated_number_of_cells
-            # Store in the project metadata
-            project.info['number_of_cells'] = number_of_cells
-            project.info.save()
-            return 0
-        except Exception as ex:
-            logger.warning("Unable to set cell count from data in "
-                           "%s: %s" % (count_dir,ex))
-    # Reached the end without setting count
-    return 1
 
 def cellranger_info(path=None,name=None):
     """
