+++ auto_process_ngs/test/test_tenx_genomics.py
@@ -8,7 +8,6 @@ import os
 import shutil
 from bcftbx.mock import MockIlluminaRun
 from bcftbx.mock import RunInfoXml
-from bcftbx.utils import mkdirs
 from auto_process_ngs.analysis import AnalysisProject
 from auto_process_ngs.mock import MockBcl2fastq2Exe
 from auto_process_ngs.mock import MockCellrangerExe
@@ -763,391 +762,3 @@ PBB,CMO302,PBB
         self.assertEqual(config_csv.sample_names,['PBA','PBB'])
         self.assertEqual(config_csv.reference_data_path,
                          "/data/refdata-cellranger-gex-GRCh38-2020-A")
-
-class TestSetCellCountForProject(unittest.TestCase):
-    """
-    Tests for the 'set_cell_count_for_project' function
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestSetCellCountForProject')
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-    def _make_mock_analysis_project(self,single_cell_platform,library_type):
-        # Create a mock AnalysisProject
-        m = MockAnalysisProject('PJB',
-                                fastq_names=("PJB1_S1_L001_R1_001.fastq.gz",
-                                             "PJB1_S1_L001_R2_001.fastq.gz",),
-                                metadata={'Single cell platform':
-                                          single_cell_platform,
-                                          'Library type': library_type,})
-        m.create(top_dir=self.wd)
-        return os.path.join(self.wd,'PJB')
-    def test_set_cell_count_for_project(self):
-        """
-        set_cell_count_for_project: test for scRNA-seq
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Chromium 3'v3",
-            "scRNA-seq")
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "5.0.1",
-                                  "refdata-gex-GRCh38-2020-A",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        metrics_summary_file = os.path.join(counts_dir,
-                                            "metrics_summary.csv")
-        with open(metrics_summary_file,'wt') as fp:
-            fp.write(METRICS_SUMMARY)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-gex-GRCh38-2020-A
-Cellranger version\t5.0.1
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         2272)
-    def test_set_cell_count_for_atac_project(self):
-        """
-        set_cell_count_for_project: test for scATAC-seq
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Single Cell ATAC",
-            "scATAC-seq")
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "1.2.0",
-                                  "refdata-cellranger-atac-GRCh38-1.2.0",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        summary_file = os.path.join(counts_dir,
-                                    "summary.csv")
-        with open(summary_file,'w') as fp:
-            fp.write(ATAC_SUMMARY)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-cellranger-atac-GRCh38-1.2.0
-Cellranger version\t1.2.0
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         5682)
-    def test_set_cell_count_for_single_nuclei_atac_project(self):
-        """
-        set_cell_count_for_project: test for snATAC-seq
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Single Cell ATAC",
-            "snATAC-seq")
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "1.2.0",
-                                  "refdata-cellranger-atac-GRCh38-1.2.0",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        summary_file = os.path.join(counts_dir,
-                                            "summary.csv")
-        with open(summary_file,'w') as fp:
-            fp.write(ATAC_SUMMARY)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-cellranger-atac-GRCh38-1.2.0
-Cellranger version\t1.2.0
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         5682)
-    def test_set_cell_count_for_atac_project_2_0_0(self):
-        """
-        set_cell_count_for_project: test for scATAC-seq (Cellranger ATAC 2.0.0)
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Single Cell ATAC",
-            "scATAC-seq")
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "2.0.0",
-                                  "refdata-cellranger-atac-GRCh38-2020-A-2.0.0",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        summary_file = os.path.join(counts_dir,
-                                    "summary.csv")
-        with open(summary_file,'w') as fp:
-            fp.write(ATAC_SUMMARY_2_0_0)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-cellranger-atac-GRCh38-2020-A-2.0.0
-Cellranger version\t2.0.0
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         3582)
-    def test_set_cell_count_for_multiome_atac_project(self):
-        """
-        set_cell_count_for_project: test for single cell multiome ATAC
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Single Cell Multiome",
-            "ATAC")
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "1.0.0",
-                                  "refdata-cellranger-arc-GRCh38-2020-A",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        summary_file = os.path.join(counts_dir,
-                                    "summary.csv")
-        with open(summary_file,'w') as fp:
-            fp.write(MULTIOME_SUMMARY)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-cellranger-arc-GRCh38-2020-A
-Cellranger version\t1.0.0
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         744)
-    def test_set_cell_count_for_multiome_gex_project(self):
-        """
-        set_cell_count_for_project: test for single cell multiome GEX
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Single Cell Multiome",
-            "GEX")
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "1.0.0",
-                                  "refdata-cellranger-arc-GRCh38-2020-A",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        summary_file = os.path.join(counts_dir,
-                                    "summary.csv")
-        with open(summary_file,'w') as fp:
-            fp.write(MULTIOME_SUMMARY)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-cellranger-arc-GRCh38-2020-A
-Cellranger version\t1.0.0
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         744)
-    def test_set_cell_count_project_missing_library_type(self):
-        """
-        set_cell_count_for_project: test for scRNA-seq when library not set
-        """
-        # Set up mock project with library type not set
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Chromium 3'v3",
-            None)
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "5.0.1",
-                                  "refdata-gex-GRCh38-2020-A",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        metrics_summary_file = os.path.join(counts_dir,
-                                            "metrics_summary.csv")
-        with open(metrics_summary_file,'w') as fp:
-            fp.write(METRICS_SUMMARY)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-gex-GRCh38-2020-A
-Cellranger version\t5.0.1
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         2272)
-    def test_set_cell_count_for_project_no_subdirs(self):
-        """
-        set_cell_count_for_project: test for scRNA-seq (old-style output)
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Chromium 3'v3",
-            "scRNA-seq")
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        metrics_summary_file = os.path.join(counts_dir,
-                                            "metrics_summary.csv")
-        with open(metrics_summary_file,'w') as fp:
-            fp.write(METRICS_SUMMARY)
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         2272)
-    def test_set_cell_count_project_missing_library_type_no_subdirs(self):
-        """
-        set_cell_count_for_project: test for scRNA-seq when library not set (old-style output)
-        """
-        # Set up mock project with library type not set
-        project_dir = self._make_mock_analysis_project(
-            "10xGenomics Chromium 3'v3",
-            None)
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        metrics_summary_file = os.path.join(counts_dir,
-                                            "metrics_summary.csv")
-        with open(metrics_summary_file,'w') as fp:
-            fp.write(METRICS_SUMMARY)
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Update the cell counts
-        print("Updating number of cells")
-        set_cell_count_for_project(project_dir)
-        # Check updated cell count
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         2272)
-    def test_set_cell_count_fails_for_project_with_no_metadata(self):
-        """
-        set_cell_count_for_project: raises exception for project with no metadata
-        """
-        # Set up mock project
-        project_dir = self._make_mock_analysis_project(None,None)
-        # Add metrics_summary.csv
-        counts_dir = os.path.join(project_dir,
-                                  "qc",
-                                  "cellranger_count",
-                                  "5.0.1",
-                                  "refdata-gex-GRCh38-2020-A",
-                                  "PJB1",
-                                  "outs")
-        mkdirs(counts_dir)
-        metrics_summary_file = os.path.join(counts_dir,
-                                            "metrics_summary.csv")
-        with open(metrics_summary_file,'wt') as fp:
-            fp.write(METRICS_SUMMARY)
-        # Add QC info file
-        with open(os.path.join(project_dir,"qc","qc.info"),'wt') as fp:
-            fp.write("""Cellranger reference datasets\t/data/refdata-gex-GRCh38-2020-A
-Cellranger version\t5.0.1
-""")
-        # Check initial cell count
-        print("Checking number of cells")
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
-        # Attempting to update the cell counts should raise
-        # NotImplementedError
-        self.assertRaises(NotImplementedError,
-                          set_cell_count_for_project,
-                          project_dir)
-        # Check cell count wasn't updated
-        self.assertEqual(AnalysisProject("PJB1",
-                                         project_dir).info.number_of_cells,
-                         None)
